moved text creation queries to Text
