+++ src/domain/text.rs
@@ -1,11 +1,19 @@
-use std::error::Error;
-
+use actix_web::{web::Json, HttpResponse};
 use serde::{Deserialize, Serialize};
-use sqlx::{query, types::Json, FromRow, PgPool};
+use serde_json::json;
+use sqlx::{FromRow, Postgres, Transaction};
+
+use super::author::Author;
 
 #[derive(Debug, Serialize, Deserialize, FromRow)]
 pub struct Metadata {
     pub genre_tags: Vec<String>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CreateText {
+    pub text: Text,
+    pub author: Author,
 }
 
 #[derive(Debug, Serialize, Deserialize, FromRow)]
@@ -19,6 +27,7 @@ pub struct Text {
 }
 
 impl Text {
+    #[allow(dead_code)]
     pub fn new(
         text_type_id: i32,
         author_id: i32,
@@ -35,19 +44,85 @@ impl Text {
         }
     }
 
-    pub async fn create(text: &Text, pool: &PgPool) -> Result<(), Box<dyn Error>> {
-        let sql =
-            "INSERT INTO texts (text_type_id, author_id, title, published, metadata) VALUES ($1, $2, $3, $4, $5)";
 
-        query(sql)
-            .bind(&text.text_type_id)
-            .bind(&text.author_id)
-            .bind(&text.title)
-            .bind(&text.published)
-            .bind(Json(&text.metadata))
-            .execute(pool)
-            .await?;
-        Ok(())
+    pub async fn with_extant_author(
+        mut txn: Transaction<'_, Postgres>,
+        author: Author,
+        create_tex: Json<CreateText>,
+    ) -> HttpResponse {
+        let author_id = author.author_id.unwrap();
+
+        let text_sql = "INSERT INTO texts (text_type_id, author_id, title, published, metadata) VALUES ($1, $2, $3, $4, $5) RETURNING text_type_id, author_id, title, published, metadata";
+
+        let ser = json!(&create_tex.text.metadata);
+        let text = match sqlx::query_as::<_, Text>(text_sql)
+            .bind(&create_tex.text.text_type_id)
+            .bind(&author_id)
+            .bind(&create_tex.text.title)
+            .bind(&create_tex.text.published)
+            .bind(&ser)
+            .fetch_one(&mut *txn)
+            .await
+        {
+            Ok(t) => t,
+            Err(e) => {
+                eprintln!("{e}");
+                return HttpResponse::InternalServerError()
+                    .json("Failed to insert new text with Error: {e}");
+            }
+        };
+
+        match txn.commit().await {
+            Ok(_) => HttpResponse::Ok().json(text),
+            Err(e) => HttpResponse::InternalServerError()
+                .json(format!("Failed to commit transaction with Error: {e}")),
+        }
+    }
+
+    pub async fn with_new_author(
+        mut txn: Transaction<'_, Postgres>,
+        create_text_and_author: Json<CreateText>,
+    ) -> HttpResponse {
+        let author_sql = "INSERT INTO authors (first_name, last_name) VALUES ($1, $2) RETURNING authors.author_id, authors.first_name, authors.last_name";
+        let author = match sqlx::query_as::<_, Author>(author_sql)
+            .bind(&create_text_and_author.author.first_name)
+            .bind(&create_text_and_author.author.last_name)
+            .fetch_one(&mut *txn)
+            .await
+        {
+            Ok(a) => a,
+            Err(e) => {
+                eprintln!("{e}");
+                return HttpResponse::InternalServerError()
+                    .json("Failed to insert new author with Error: {e}");
+            }
+        };
+
+        let text_sql = "INSERT INTO texts (text_type_id, author_id, title, published, metadata) VALUES ($1, $2, $3, $4, $5) RETURNING text_type_id, author_id, title, published, metadata";
+        let json_metadata = json!(&create_text_and_author.text.metadata);
+
+        let text = match sqlx::query_as::<_, Text>(text_sql)
+            .bind(&create_text_and_author.text.text_type_id)
+            .bind(&author.author_id)
+            .bind(&create_text_and_author.text.title)
+            .bind(&create_text_and_author.text.published)
+            .bind(&json_metadata)
+            .fetch_one(&mut *txn)
+            .await
+        {
+            Ok(text) => text,
+            Err(e) => {
+                eprintln!("{e}");
+                return HttpResponse::InternalServerError()
+                    .json(format!("Failed to insert new text with Error: {e}"));
+            }
+        };
+
+        match txn.commit().await {
+            Ok(_) => HttpResponse::Ok().json(text),
+            Err(e) => HttpResponse::InternalServerError()
+                .json(format!("Failed to commit transaction with Error: {e}")),
+        }
     }
 }
 
@@ -56,10 +131,5 @@ pub struct TitleWithAuthor {
     title: String,
     first_name: String,
     last_name: String,
-}
-
-impl TitleWithAuthor {
-    pub fn new(title: String, first_name: String, last_name: String) -> Self {
-        Self { title, first_name, last_name }
-    }
+    published: i32,
 }
